fix(#4): run command use `--config` flag if passed
This fixes the run command if the user uses a the configuration file on a different location than `$HOME`.

Note: This doesn't fix when installing vale, because on that case it will be needed to generate a `vale.ini` file,
without that `vale` complains about not being able to find a configruationfile.